+++ sql/api/src/main/scala/org/apache/spark/sql/types/StructField.scala
@@ -149,6 +149,7 @@ case class StructField(
 
   /**
    * Return the default value of this StructField.
+   * This is used for storing the default value of a function parameter.
    */
   private[sql] def getDefault(): Option[String] = {
     if (metadata.contains("default")) {
